se aplica estilos a Pedido
