--- src/Estilos/VistaCocina.scss
@@ -1,6 +1,0 @@
-.VistaCocina  {
-    display: flex;
-    flex-direction: row;
-    background: #252836;
-    color: aliceblue;
-}
