--- src/Estilos/VistaMesero.scss
@@ -1,8 +1,0 @@
-.VistaMesero, .VistaCocina ,.Pedido {
-    display: flex;
-    flex-direction: row;
-    background: #252836;
-    color: aliceblue;
-    width: 1194px;
-    height: 834px;
-}
