+++ src/componente/CabeceraOrden.jsx
@@ -1,7 +1,7 @@
 import React, { useContext, useEffect } from 'react';
-import AppContext from '../context/AppContext';
+import AppContext from '../contexto/AppContext';
 import { obtenerOrdenes } from '../firebase/firebaseStore';
-import '../estilos/CabeceraOrden.scss';
+import '../estilo/CabeceraOrden.scss';
 
 export default function CabeceraOrden() {
   const { orden, obtenerNumeroOrden } = useContext(AppContext);
