+++ src/contenedor/Pedidos.jsx
@@ -1,12 +1,13 @@
 import React, { useState, useEffect } from 'react';
 import { obtenerOrdenesFiltradas } from '../firebase/firebaseStore';
-import CabeceraUsuario from '../Componentes/CabeceraUsuario';
-import BarraSecundaria from '../Componentes/BarraSecundaria';
-import PedidoPendiente from '../Componentes/PedidoPendiente';
-import '../estilos/Notificaciones.scss';
+import BarraPrincipal from '../componente/BarraPrincipal';
+import CabeceraUsuario from '../componente/CabeceraUsuario';
+import BarraSecundaria from '../componente/BarraSecundaria';
+import PedidoPendiente from '../componente/PedidoPendiente';
+import '../estilo/Notificaciones.scss';
 
 export default function Pedidos() {
-  const [subcategoria, setSubcategoria] = useState('Recibidos');
+  const [subcategoria, setSubcategoria] = useState('Por preparar');
 
   // Funcion obtener productos de Firestore
   const bdOrdenes = async () => {
@@ -20,7 +21,6 @@ export default function Pedidos() {
 
   // Estado del arreglo de productos
   const [arregloOrdenes, setArregloOrdenes] = useState([]);
-
   // Efecto para que se actualice los productos cuando cambie la subcategoria
   useEffect(() => bdOrdenes().then((arr) => {
     console.log(arr);
@@ -28,10 +28,12 @@ export default function Pedidos() {
   }), [subcategoria]);
   return (
     <div className="Pedidos">
+      <BarraPrincipal vista="Pedidos" />
       <CabeceraUsuario />
       <BarraSecundaria cat="Pedidos" setSubcategoria={setSubcategoria} />
-      <div className="contenedorPedidos" />
-      {arregloOrdenes.map((orden) => <PedidoPendiente key={orden.numeroOrden} orden={orden} subcategoria={subcategoria} />)}
+      <div className="contenedorPedidos">
+        {arregloOrdenes.map((orden) => <PedidoPendiente key={orden.numeroOrden} orden={orden} subcategoria={subcategoria} />)}
+      </div>
     </div>
   );
 }
