+++ src/estilo/BarraSecundaria.scss
@@ -9,7 +9,7 @@ $color_anular: #FF7CA3;
 
 .BarraSecundaria{
   margin-top: 15px;
-    width: auto;
+    width: 100%;
 }
 
 ul{
