+++ src/estilo/CabeceraUsuario.scss
@@ -8,6 +8,7 @@
     align-items: center;
     margin-left: 2rem;
     margin-right: 2rem;
+    color: aliceblue;
 }
 
 .fecha{
