+++ src/estilo/DetalleOrden.scss
