+++ src/pagina/InicioSesion.jsx
@@ -2,11 +2,11 @@ import React, { useRef, useContext } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faUserCircle } from '@fortawesome/free-solid-svg-icons';
-import AppContext from '../context/AppContext';
+import AppContext from '../contexto/AppContext';
 import { iniciarSesion } from '../firebase/firebaseAuth';
 import { obtenerUsuario } from '../firebase/firebaseStore';
-import logo from '../imagenes/logoHamburguesa.svg';
-import '../estilos/InicioSesion.scss';
+import logo from '../imagen/logoHamburguesa.svg';
+import '../estilo/InicioSesion.scss';
 
 export default function InicioSesion() {
   const { obtenerNombreMesero } = useContext(AppContext);
@@ -26,7 +26,7 @@ export default function InicioSesion() {
           obtenerNombreMesero(docu.nombre);
           console.log(docu.rol);
           if (docu.rol === 'Mesero') navigate('/VistaMesero/Desayuno');
-          else navigate('/VistaCocina');
+          else navigate('/Pedidos');
         });
       })
       .catch((e) => console.log('Hubo un error en la authenticacion', e));
